test: nil types do not cause panic
